Move headless-service-specific functions to cluster_watcher_headless.go
This just unclutters a bit `cluster_watcher.go` by moving all the
functions specific to headless services into the new
`cluster_watch_headless.go` file.
